break workflow out into separate modules, so multiple workflows can operate over the inbox
