+++ invenio_notify/constants.py
@@ -11,7 +11,6 @@ TYPE_ENDORSEMENT = 'coar-notify:EndorsementAction'
 TYPE_TENTATIVE_ACCEPT = 'TentativeAccept'
 TYPE_REJECT = 'Reject'
 TYPE_TENTATIVE_REJECT = 'TentativeReject'
-SUPPORTED_TYPES = [TYPE_REVIEW, TYPE_ENDORSEMENT, TYPE_TENTATIVE_ACCEPT, TYPE_REJECT, TYPE_TENTATIVE_REJECT]
 """ List of supported notification types that can be processed by the notify  """
 
 # Workflow status constants
